Abstract out useful things into a separate class we can use for more than one kind of test.
