+++ app/src/androidTest/java/com/zacharytamas/often/utils/dates/NextAvailableAtTest.java
@@ -1,11 +1,7 @@
 package com.zacharytamas.often.utils.dates;
 
-import android.test.InstrumentationTestCase;
-
-import com.zacharytamas.often.models.Habit;
 import com.zacharytamas.often.models.RepeatType;
 import com.zacharytamas.often.models.RepeatUnit;
-import com.zacharytamas.often.utils.Dates;
 
 import org.joda.time.DateTime;
 
@@ -18,34 +14,7 @@ import static com.zacharytamas.often.utils.Dates.nextAvailableAt;
 /**
  * Created by zacharytamas on 12/12/14.
  */
-public class NextAvailableAtTest extends InstrumentationTestCase {
-
-    Habit habit;
-    Date now;
-
-    @Override
-    protected void setUp() throws Exception {
-        super.setUp();
-
-        habit = new Habit();
-        habit.setTitle("Brush teeth");
-
-        now = getDate(1992, 0, 31);
-    }
-
-    private void assertSameDay(Date d1, Date d2) {
-        DateTime dateTime1 = new DateTime(d1);
-        DateTime dateTime2 = new DateTime(d2);
-
-        assertEquals(dateTime1.withTimeAtStartOfDay(),
-                     dateTime2.withTimeAtStartOfDay());
-    }
-
-    private Date getDate(int year, int month, int date) {
-        Calendar calendar = GregorianCalendar.getInstance();
-        calendar.set(year, month, date, 0, 0, 0);
-        return calendar.getTime();
-    }
+public class NextAvailableAtTest extends DatesTest {
 
     public void test_repeatType_periodically() {
